requirements.txt
wheel
Flask
Flask-SQLAlchemy
python-dotenv
pytest
gunicorn
flasgger
requests
flask-cors
pymongo[srv]